Added `yrText`
Introducing a new member to `yrKit`: yrText!

This version is not the complete version of `yrText`. This is only a web app designed for `yrHost`. The future desktop version of `yrText` will have much more features than just saving. Also going to be working on making this web app friendly for mobile users. Or maybe I just won't let mobile users edit files, since `yrText` is originally meant to be a desktop app. Who knows ¯\_(ツ)_/¯