TESTING: using "packs" concept instead of awaiting chunks
I recently came up with a new concept to upload files that I call "packs". Basically, a pack is a folder of folders that include 3MB chunks of data in a `.part` file.

A folder in a pack is called an "entry", which represents a file. In each entry, there is a configuration file (with JSON, ofc) and chunks of a file that are numbered and include a 3MB slice of a file. The configuration file just includes the relative path of the entry so that this packs concept also supports sending in folders.

This packs concept can help, so that we don't have to wait to upload each individual chunk to a file. We just upload all chunks, and wait for all of them to finish together. After that, we can join all of the chunks in an entry and then we can delete the pack.

Lmk if there are any flaws about this. I've already fixed one, which was that if I had sent too many requests at once, it would be too much for the browser and the server if there are multiple people uploading a big file. So I capped the concurrent requests to be 5 instead of unlimited. Will try to make a setting to customize this cap in the `config.json`.

This concept still isn't tested yet!! I'm only uploading this so that I can test on my test laptop.